x/run: feed the commands to remote host over ssh stdin
Previously we'd bother invoking $SHELL -c and passing the escaped
command line as an argument, which was complex to implement and
inflexible.

We can instead construct a proper script and feed it over pipe to
remote shell, which is easy to implement and is far more flexible.
It has also provided an opportunity to enable the following:

	cat <<. | x/run
	echo not bad
	.